oakenc: refactor into several files
start work on anaglyph
start using cv2
