+++ src/ros/oak/shell.nix
@@ -6,6 +6,7 @@ in pkgs.mkShell {
     (pkgs.python3.withPackages (python-pkgs: with python-pkgs; [
       numpy
       pillow
+      opencv4
     ]))
   ];
 }
